Refactoring structure of the project
