--- src/main/java/org/javabasictraining/openedApproach/Knight.java
@@ -1,14 +1,0 @@
-package org.javabasictraining.openedApproach;
-
-public class Knight extends Warrior {
-    private static final int ATTACK = 7;
-
-    public Knight() {
-        super(INITIAL_HEALTH, ATTACK);
-    }
-
-    @Override
-    public int getAttack() {
-        return ATTACK;
-    }
-}
