--- src/main/java/org/javabasictraining/openedApproach/Unit.java
@@ -1,6 +1,0 @@
-package org.javabasictraining.openedApproach;
-
-public interface Unit {
-
-    void joinArmy(Army army);
-}
