+++ src/test/java/org/study/warriors/ArmyBasicFunctionTestSuite.java
@@ -1,15 +1,15 @@
-package org.javabasictraining.encapsulatedApproach;
+package org.study.warriors;
 
-import org.javabasictraining.encapsulatedApproach.Army;
-import org.javabasictraining.encapsulatedApproach.Unit;
-import org.javabasictraining.encapsulatedApproach.Warrior;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.study.warriors.model.Army;
+import org.study.warriors.model.Unit;
+import org.study.warriors.model.Warrior;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class ArmyTestSuite {
+class ArmyBasicFunctionTestSuite {
 
     private final Army army = new Army();
